requirements.txt
arabic-reshaper==3.0.0
asgiref==3.9.1
asn1crypto==1.5.1
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
contourpy==1.3.3
cryptography==45.0.5
cssselect2==0.8.0
cycler==0.12.1
diff-match-patch==20241021
Django==5.2.4
django-import-export==4.3.9
fonttools==4.59.0
gunicorn==23.0.0
html5lib==1.1
idna==3.10
kiwisolver==1.4.8
lxml==6.0.0
matplotlib==3.10.5
numpy==2.3.2
oscrypto==1.3.0
packaging==25.0
pandas==2.3.1
pillow==11.3.0
pycparser==2.22
pyHanko==0.29.1
pyhanko-certvalidator==0.27.0
pyparsing==3.2.3
pypdf==5.9.0
python-bidi==0.6.6
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
reportlab==4.4.3
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
svglib==1.5.1
tablib==3.8.0
tinycss2==1.4.0
tzdata==2025.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.5.0
webencodings==0.5.1
whitenoise==6.9.0
xhtml2pdf==0.2.17
